Refactor: Camera, GameConsole to module
